Add files via upload
pdbview will graph the positions of resiudes in a pdb on an interactive plot with lines drawn between the residues that are in contact, either by mdtraj selections or by smog-server .contacts file. the graph can be moved and rotated and residues and their respective contacts can be toggled on and off.