GMPP V 1.0.0
Added landscape generation. Now the GMPP is capable of running and completely analyzing Go-Model simulations with files from the SMOG-server. The final outputs will consist of: Native Contacts for all the simulations, RMSD for the folding temperature, and Radius of Gyration all 3 of which will be used with various starting references to generate landscapes with the dimensionless free energies calculated via the WHAM algorithm. 

I'm working on a user manual for the downloadable version of the GMPP, the readme file has a complete rundown of the usage.